Add support for new department and selected class seed data, update database schema for class IDs to use BigInteger, and enhance filtering options in the details page.

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, DateTime, Boolean, ForeignKey, DECIMAL, Date
+from sqlalchemy import Column, Integer, BigInteger, String, Text, DateTime, Boolean, ForeignKey, DECIMAL, Date
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.db.base import Base
@@ -54,7 +54,7 @@ class Student(Base):
 class Class(Base):
     __tablename__ = "classes"
     
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     name = Column(String(64), nullable=False, index=True)
     credits = Column(Integer, nullable=False)
     requiredOrElectiveCourse = Column(String(64), nullable=False, index=True)
@@ -66,7 +66,7 @@ class Class(Base):
 class SelectedClass(Base):
     __tablename__ = "selected_classes"
     
-    classid = Column(Integer, ForeignKey('classes.id'), primary_key=True, index=True)
+    classid = Column(BigInteger, ForeignKey('classes.id'), primary_key=True, index=True)
     studentid = Column(Integer, ForeignKey('students.id'), primary_key=True, index=True)
     ispassed = Column(Boolean, default=False)
     score = Column(DECIMAL(10, 2), default=0.00)
